bring expand_config to `common`
